refactor InfoMessage test
