+++ src/components/atoms/InfoMessage.unit.ts
@@ -1,0 +1,34 @@
+import { mount } from "@vue/test-utils";
+import { createTestingVuetify } from "@@/tests/test-utils/setup";
+import InfoMessage from "./InfoMessage.vue";
+
+describe("@/components/atoms/InfoMessage", () => {
+	const expectedMessage = "This is some useful information";
+	const setup = (type?: string) => {
+		const wrapper = mount(InfoMessage, {
+			global: {
+				plugins: [createTestingVuetify()],
+			},
+			props: { message: expectedMessage, type },
+		});
+
+		return wrapper;
+	};
+
+	it("shows the message", async () => {
+		const defaultType = "bc-info";
+		const wrapper = setup();
+
+		expect(wrapper.element.className).toContain(defaultType);
+		expect(wrapper.find("div.message").element.textContent?.trim()).toBe(
+			expectedMessage
+		);
+	});
+
+	it("has correct styling class", async () => {
+		const expectedClass = "bc-error";
+		const wrapper = setup(expectedClass);
+
+		expect(wrapper.element.className).toContain(expectedClass);
+	});
+});
